pyproject.toml
[project]
name = "fihrist"
version = "0.1.0"
description = "A multi-tenant CMS"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "email-validator>=2.3.0",
    "flask>=3.0.3",
    "flask-babel>=4.0.0",
    "flask-login>=0.6.3",
    "flask-mail>=0.10.0",
    "flask-migrate>=4.1.0",
    "flask-sqlalchemy>=3.1.1",
    "flask-wtf>=1.2.1",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "aiosmtpd>=1.4.6",
    "faker>=35.2.2",
]
